refactor: move rendering utils into sections package
Move format_total_seconds, parse_distance_km, days_ago, weeks_and_days_until,
and combine_sections from coach/utils.py into coach/reasoning/coach/sections/utils.py
— closer to the sections that use them. coach/utils.py retains only ingestion/domain
utilities (parse_utc_datetime, parse_private_notes_activity_summary).
